+++ CHANGES.md
@@ -1,3 +1,4 @@
+- 20240127 remove instqueue, now replaced by "instpackage queue-only"
 - 20221115 honor all groups of qmailr user when spawning qmail-rspawn
 - 20221019 code: netstring parsing in qmail-qmtpd & qmail-qmqpd now
            checks that the netstring length is actually a number.